// enum/enum.ts
export enum CartWarnings {
    CompleteHeaderText = 'Thank you for your order!',
    CompleteOrderText = 'Your order has been dispatched, ' +
        'and will arrive just as fast as the pony can get there!',
}

export enum FooterCopyTexts {
    FooterCopyrightText = '© 2025 Sauce Labs. All Rights Reserved. ' +
        'Terms of Service | Privacy Policy',
}

export enum LoginPageWarnings {
    LoginError = 'Epic sadface: Username and password ' +
        'do not match any user in this service',
}

export enum Selectors {
    InventoryItemName = 'inventory-item-name',
    InventoryItemDesc = 'inventory-item-desc',
    InventoryItemPrice = 'inventory-item-price',
    PaymentInfoLabel = 'payment-info-label',
    PaymentInfoValue = 'payment-info-value',
    ShippingInfoLabel = 'shipping-info-label',
    ShippingInfoValue = 'shipping-info-value',
    SubtotalLabel = 'subtotal-label',
    TaxLabel = 'tax-label',
    TotalLabel = 'total-label',
    ItemQuantity = 'item-quantity',
    TextHeader = 'complete-header',
    TextComplete = 'complete-text',
    SocialTwitter = 'social-twitter',
    SocialFacebook = 'social-facebook',
    SocialLinkedin = 'social-linkedin',
    FooterCopyText = 'footer-copy',
    ShoppingCartLink = 'shopping-cart-link',
    AddToCartItem = 'add-to-cart-sauce-labs-backpack',
    AddToCartApprove = 'add-to-cart',
    AboutSidebar = 'about-sidebar-link',
    Title = 'title',
}